apps/core/tests/test_config.py
"""
Tests for the centralized configuration system.
"""

from django.test import TestCase, override_settings
from django.contrib.auth import get_user_model
from pydantic import ValidationError
import pytest

from apps.core.config import (
    SearchConfig, APIConfig, ProcessingConfig, SystemConfig,
    get_config, reload_config
)
from apps.core.models import Configuration

User = get_user_model()


class ConfigModelTests(TestCase):
    """Tests for Configuration model."""
    
    def setUp(self):
        self.user = User.objects.create_user(
            username='testuser',
            email='test@example.com',
            password='testpass123'
        )
    
    def test_create_configuration(self):
        """Test creating a configuration entry."""
        config = Configuration.objects.create(
            key='test_key',
            value={'test': 'value'},
            description='Test configuration',
            updated_by=self.user
        )
        
        self.assertEqual(config.key, 'test_key')
        self.assertEqual(config.value, {'test': 'value'})
        self.assertEqual(config.updated_by, self.user)
    
    def test_get_config_existing(self):
        """Test getting existing configuration."""
        Configuration.objects.create(
            key='test_key',
            value={'test': 'value'}
        )
        
        value = Configuration.get_config('test_key')
        self.assertEqual(value, {'test': 'value'})
    
    def test_get_config_default(self):
        """Test getting configuration with default."""
        value = Configuration.get_config('nonexistent', default='default_value')
        self.assertEqual(value, 'default_value')
    
    def test_set_config(self):
        """Test setting configuration."""
        config = Configuration.set_config(
            'test_key',
            {'new': 'value'},
            description='Updated config',
            user=self.user
        )
        
        self.assertEqual(config.key, 'test_key')
        self.assertEqual(config.value, {'new': 'value'})
        self.assertEqual(config.description, 'Updated config')
        self.assertEqual(config.updated_by, self.user)
    
    def test_delete_config(self):
        """Test deleting configuration."""
        Configuration.objects.create(key='test_key', value='test')
        
        deleted = Configuration.delete_config('test_key')
        self.assertTrue(deleted)
        
        # Try to delete non-existent
        deleted = Configuration.delete_config('test_key')
        self.assertFalse(deleted)
    
    def test_bulk_update(self):
        """Test bulk updating configurations."""
        configs = {
            'key1': {'value': 1},
            'key2': {'value': 2},
            'key3': {'value': 3}
        }
        
        created, updated = Configuration.bulk_update(configs, self.user)
        
        self.assertEqual(created, 3)
        self.assertEqual(updated, 0)
        
        # Update existing
        configs['key1'] = {'value': 'updated'}
        created, updated = Configuration.bulk_update(configs, self.user)
        
        self.assertEqual(created, 0)
        self.assertEqual(updated, 3)


class PydanticConfigTests(TestCase):
    """Tests for Pydantic configuration models."""
    
    def test_search_config_validation(self):
        """Test SearchConfig validation."""
        # Valid config
        config = SearchConfig(
            default_num_results=50,
            default_language='fr',
            default_file_types=['pdf', 'doc']
        )
        self.assertEqual(config.default_num_results, 50)
        
        # Invalid num_results (too high)
        with pytest.raises(ValidationError):
            SearchConfig(default_num_results=200)
        
        # Invalid language format
        with pytest.raises(ValidationError):
            SearchConfig(default_language='english')
    
    def test_api_config_validation(self):
        """Test APIConfig validation."""
        # Valid config
        config = APIConfig(
            serper_timeout=60,
            rate_limit_per_minute=200,
            max_retries=5
        )
        self.assertEqual(config.serper_timeout, 60)
        
        # Invalid timeout (too high)
        with pytest.raises(ValidationError):
            APIConfig(serper_timeout=150)
    
    def test_processing_config_validation(self):
        """Test ProcessingConfig validation."""
        # Valid config
        config = ProcessingConfig(
            batch_size=100,
            cache_ttl=7200,
            duplicate_threshold=0.9
        )
        self.assertEqual(config.batch_size, 100)
        
        # Invalid threshold (too high)
        with pytest.raises(ValidationError):
            ProcessingConfig(duplicate_threshold=1.5)
    
    @override_settings(THESIS_GREY_CONFIG={
        'search': {'default_num_results': 50},
        'api': {'serper_timeout': 45}
    })
    def test_system_config_load_from_settings(self):
        """Test loading SystemConfig from Django settings."""
        config = SystemConfig.load_from_settings()
        
        self.assertEqual(config.search.default_num_results, 50)
        self.assertEqual(config.api.serper_timeout, 45)
        # Check defaults are used for unspecified values
        self.assertEqual(config.processing.batch_size, 50)
    
    def test_system_config_save_to_db(self):
        """Test saving SystemConfig to database."""
        config = SystemConfig.load_from_settings()
        config.save_to_db()
        
        # Check it was saved
        db_config = Configuration.get_config('system_config')
        self.assertIsNotNone(db_config)
        self.assertEqual(db_config['search']['default_num_results'], 100)
    
    @override_settings(THESIS_GREY_CONFIG={'search': {'default_num_results': 75}})
    def test_config_singleton(self):
        """Test configuration singleton behavior."""
        config1 = get_config()
        config2 = get_config()
        
        # Should be the same instance
        self.assertIs(config1, config2)
        
        # Reload should create new instance
        config3 = reload_config()
        self.assertIsNot(config1, config3)
        
        # But get_config should now return the new instance
        config4 = get_config()
        self.assertIs(config3, config4)
    
    def test_config_with_database_override(self):
        """Test configuration with database override."""
        # Set initial config in database
        Configuration.set_config('system_config', {
            'search': {'default_num_results': 25},
            'api': {'serper_timeout': 15}
        })
        
        # Load config - should merge with settings
        config = SystemConfig.load_from_settings()
        
        # Database values should override settings
        self.assertEqual(config.search.default_num_results, 25)
        self.assertEqual(config.api.serper_timeout, 15)
        # Other values should use defaults
        self.assertEqual(config.processing.batch_size, 50)
    
    @override_settings(THESIS_GREY_CONFIG={'search': {'default_num_results': 80}})
    def test_constance_integration(self):
        """Test integration with Django Constance."""
        # This test would need Constance to be available
        try:
            from constance import config as constance_config
            # If constance is available, test the integration
            config = SystemConfig.load_from_settings()
            # Values should come from settings or constance
            self.assertIsInstance(config.search.default_num_results, int)
        except ImportError:
            # Skip if constance not available
            self.skipTest("Django Constance not available")
    
    def test_config_file_types_parsing(self):
        """Test parsing of file types from string."""
        config_data = {
            'search': {
                'default_file_types': 'pdf,doc,docx'
            }
        }
        Configuration.set_config('system_config', config_data)
        
        config = SystemConfig.load_from_settings()
        # Should handle string parsing in the actual implementation
        self.assertIsInstance(config.search.default_file_types, list)